requirements.txt
dnspython
plotext